fix(HTD-60): simplify rich text
