+++ src/app/privacy-policy/content.mdx
@@ -6,9 +6,7 @@ This policy applies to all information collected through use of our website, HTT
 
 When you use our Services, you trust us with your personal information. We take your privacy very seriously. In this privacy notice, we describe our privacy policy. We seek to explain to you in the clearest way possible what information we collect, how we use it and what rights you have in relation to it. We hope you take some time to read through it carefully, as it is important. If there are any terms in this privacy policy that you do not agree with, please discontinue use of the Services.
 
-<Spacer />
-
-## What information do we collect?
+<Heading2 $margin={48}>What information do we collect?</Heading2>
 
 Personal information you disclose to us
 
@@ -30,9 +28,7 @@ We automatically collect certain information when you visit, use or navigate the
 
 This information is primarily needed to maintain the security and operation of our Services, and for our internal analytics and reporting purposes. We will make best efforts to ensure this information is anonymous wherever possible.
 
-<Spacer />
-
-## How do we use your information?
+<Heading2 $margin={48}>How do we use your information?</Heading2>
 
 In short: we process your information for purposes based on legitimate business interests, the fulfillment of our contract with you, compliance with our legal obligations, and/or your consent.
 
@@ -58,9 +54,7 @@ We use the information we collect or receive:
 
 - For other business purposes. We may use your information for other Business Purposes, such as data analysis, identifying usage trends, determining the effectiveness of our promotional campaigns and to evaluate and improve our Services and your experience.
 
-<Spacer />
-
-## Will your information be shared with anyone?
+<Heading2 $margin={48}>Will your information be shared with anyone?</Heading2>
 
 In short: we only share information with your consent, to comply with laws, to protect your rights, or to fulfill business obligations.
 
@@ -84,9 +78,7 @@ More specifically, we may need to process your data or share your personal infor
 
 - Business Transfers. We may share or transfer your information in connection with, or during negotiations of, any merger, sale of company assets, financing, or acquisition of all or a portion of our business to another company.
 
-<Spacer />
-
-### With whom could my data be shared?
+<Heading3 $margin={48}>With whom could my data be shared?</Heading3>
 
 We only share and disclose your information with the following third parties. We have categorized each party so that you may be easily understand the purpose of our data collection and processing practices. If we have processed your data based on your consent and you wish to revoke your consent, please contact us.
 
@@ -106,31 +98,23 @@ We only share and disclose your information with the following third parties. We
 
 - Sentry: for error reporting
 
-<Spacer />
-
-## How long do we keep your information?
+<Heading2 $margin={48}>How long do we keep your information?</Heading2>
 
 We will only keep your personal information for as long as it is necessary for the purposes set out in this privacy policy, unless a longer retention period is required or permitted by law (such as tax, accounting or other legal requirements). No purpose in this policy will require us keeping your personal information for longer than 90 days past the termination of the user's account.
 
 When we have no ongoing legitimate business need to process your personal information, we will either delete or anonymize it, or, if this is not possible (for example, because your personal information has been stored in backup archives), then we will securely store your personal information and isolate it from any further processing until deletion is possible.
 
-<Spacer />
-
-## How do we keep your information safe?
+<Heading2 $margin={48}>How do we keep your information safe?</Heading2>
 
 We have implemented appropriate technical and organisational security measures designed to protect the security of any personal information we process. However, please also remember that we cannot guarantee that the internet itself is 100% secure. Although we will do our best to protect your personal information, transmission of personal information to and from our Services is at your own risk. You should only access the services within a secure environment.
 
-<Spacer />
-
-## Do we collect information from minors?
+<Heading2 $margin={48}>Do we collect information from minors?</Heading2>
 
 We do not knowingly solicit data from or market to children under 13 years of age. By using the Services, you represent that you are at least 13 or that you are the parent or guardian of such a minor and consent to such minor dependent’s use of the Services.
 
 If we learn that personal information from users less than 18 years of age has been collected, we will deactivate the account and take reasonable measures to promptly delete such data from our records. If you become aware of any data we have collected from children under age 13, please contact us using the details below.
 
-<Spacer />
-
-## What are my privacy rights?
+<Heading2 $margin={48}>What are my privacy rights?</Heading2>
 
 In some regions (like the European Economic Area), you have certain rights under applicable data protection laws. These may include the right (i) to request access and obtain a copy of your personal information, (ii) to request rectification or erasure; (iii) to restrict the processing of your personal information; and (iv) if applicable, to data portability. In certain circumstances, you may also have the right to object to the processing of your personal information. To make such a request, please use the contact details provided below. We will consider and act upon any request in accordance with applicable data protection laws.
