+++ src/app/term-of-services/content.mdx
@@ -1,15 +1,11 @@
 <Heading2 id="example-1">Terms of Service ("Terms")</Heading2>
-Please read these Terms of Service ("Terms", "Terms of Service", "Terms of Use") carefully before using the httptoolkit.com
-website, any HTTP Tookit user accounts, or the published HTTP Toolkit applications (the "Services") operated by Timothy Perry,
-trading as HTTP Toolkit ("us", "we", or "our").
+Please read these Terms of Service ("Terms", "Terms of Service", "Terms of Use") carefully before using the httptoolkit.com website, any HTTP Tookit user accounts, or the published HTTP Toolkit applications (the "Services") operated by Timothy Perry, trading as HTTP Toolkit ("us", "we", or "our").
 
 Your access to and use of the Services is conditioned on your acceptance of and compliance with these Terms. These Terms apply to all visitors, users and others who access or use the Services.
 
 By accessing or using the Services you agree to be bound by these Terms. If you disagree with any part of the terms please do not access the Services.
 
-<Spacer $spacing="policies" />
-
-### Use License
+<Heading3 $margin={48}>Use License</Heading3>
 
 All materials provided through the Services are the copyright of HTTP Toolkit, except where explicitly stated. The name, logos and trademarks of HTTP Toolkit may not be used outside of the terms provided without explicit approval from HTTP Toolkit.
 
@@ -19,9 +15,7 @@ The HTTP Toolkit application downloadable from httptoolkit.com may be used under
 
 Other product and company names, brands or logos used within the Services are the property of their respective owners, and are used for identification purposes only. Use of these marks does not imply endorsement. Unless you inform us otherwise, you allow HTTP Toolkit to use your company name and logo, exclusively to identify your organization as a user of HTTP Toolkit, without implying any further endorsement or relationship. To opt out of this, please contact us and we'll ensure your name and logo is never used in any future HTTP Toolkit materials, and is removed from any in which they currently appear.
 
-<Spacer $spacing="policies" />
-
-<Heading3 id="example-2">Warranties and Liability</Heading3>
+<Heading3 id="example-2" $margin={48}>Warranties and Liability</Heading3>
 
 The Services are provided on an "as is", "as available" basis, without representations of warranties of any kind, to the fullest extent permitted by law. We specifically dislaim all warranties and conditions, expressed or implied. Furthermore, we do not make any representations concerning the accuracy or reliability of the use of the materials on the Services or otherwise relating to such materials or any sites linked to these Services. HTTP Toolkit shall not have any liability or responsibility for errors or omissions in the Services, errors or omissions caused by your use of the Services, or any other damages you may incur in connection with the Services.
 
@@ -29,15 +23,11 @@ HTTP Toolkit or our suppliers will not be hold accountable for any damages that 
 
 In some jurisdictions there may be limitations on these limitations, and they may not fully apply. In this case, all other remaining limitations of liability and warranty will continue to apply in full.
 
-<Spacer $spacing="policies" />
-
-### Purchases
+<Heading3 id="example-2" $margin={48}>Purchases</Heading3>
 
 Our order process is conducted by our online resellers Paddle.com and PayPro Global (depending on the region & transaction details). Paddle.com and PayPro Global are the Merchants of Record for all our orders, and they handle all customer service inquiries, returns & billing support.
 
-<Spacer $spacing="policies" />
-
-### User Accounts
+<Heading3 $margin={48}>User Accounts</Heading3>
 
 When you create an account with us, you must provide us information that is accurate, complete, and current at all times. Failure to do so constitutes a breach of the Terms, which may result in immediate termination of your account on our Services.
 
@@ -49,10 +39,8 @@ Both Professional and Team accounts grant the individuals included in the licens
 
 Account subscriptions may be cancelled at any time, either from the Settings page within the tool (for Professional accounts), from the team management dashboard (for Team accounts), or by contacting us or the reseller used (either Paddle.com or PayPro Global) directly. Once cancelled, the subscription will remain active until the end of the previously purchased period, and will then expire without further renewal.
 
-Customer support in using HTTP Toolkit may be provided at our discretion, but neither license guarantees such support or any specific level or response time of customer service, unless agreed separately.
+Customer support in using HTTP Toolkit may be provided at our discretion, but neither license guarantees such support or any specific level or response time of customer service, unless agreevd separately.
 
 You are responsible for safeguarding your HTTP Toolkit user account, and for any activities or actions taken using that account.
 
 You agree not to share your user account with any third party. You must notify us immediately upon becoming aware of any breach of security or unauthorized use of your account.
-
-<Spacer $spacing="policies" />
