+++ src/components/sections/rich-text/components/default.types.ts
@@ -1,0 +1,8 @@
+export interface SpacingShape {
+  mobile?: string;
+  desktop: string;
+}
+
+export interface StyledHeadingProps {
+  $margin?: number;
+}
