+++ src/components/sections/rich-text/components/index.tsx
@@ -1,22 +1,22 @@
-import { StyledBiggerText, StyledHeading, StyledLink, StyledSectionBreak, StyledText } from './default.styles';
-import type { StyledSpacerProps } from './default.types';
+import { StyledBiggerText, StyledHeading, StyledLink, StyledText, StyledUL } from './default.styles';
+import type { StyledHeadingProps } from './default.types';
 
 import { Accordion } from '@/components/modules/accordion';
 import type { AccordionProps } from '@/components/modules/accordion/accordion.types';
 import { CTABox } from '@/components/modules/cta-box';
 import type { CTABoxProps } from '@/components/modules/cta-box/cta-box.types';
 
-const Heading2 = ({ children, id }: Component) => {
+const Heading2 = ({ children, id, $margin }: Component<StyledHeadingProps>) => {
   return (
-    <StyledHeading forwardedAs="h2" fontSize="m" color="lightGrey" id={id}>
+    <StyledHeading $margin={$margin} forwardedAs="h2" fontSize="m" color="lightGrey" id={id}>
       {children}
     </StyledHeading>
   );
 };
 
-const Heading3 = ({ children, id }: Component) => {
+const Heading3 = ({ children, id, $margin }: Component<StyledHeadingProps>) => {
   return (
-    <StyledHeading forwardedAs="h3" fontSize="s" color="lightGrey" id={id}>
+    <StyledHeading $margin={$margin} forwardedAs="h3" fontSize="s" color="lightGrey" id={id}>
       {children}
     </StyledHeading>
   );
@@ -45,13 +45,13 @@ export const defaultComponents = {
       </StyledBiggerText>
     );
   },
-  Spacer({ $spacing }: StyledSpacerProps) {
-    return <StyledSectionBreak $spacing={$spacing} />;
-  },
   Accordion(props: AccordionProps) {
     return <Accordion {...props} />;
   },
   CTABox(props: CTABoxProps) {
     return <CTABox {...props} />;
+  },
+  ul({ children }: Component) {
+    return <StyledUL>{children}</StyledUL>;
   },
 };
