+++ src/components/sections/rich-text/rich-text.types.ts
@@ -1,6 +1,6 @@
-import type { MDXProps } from 'mdx/types';
+import type { MDXComponents, MDXProps } from 'mdx/types';
 
 export interface RichTextProps {
   content(props: MDXProps): JSX.Element;
-  theme?: 'policies' | 'default';
+  components?: MDXComponents;
 }
