--- src/components/sections/rich-text/themes/default/default.types.ts
@@ -1,8 +1,0 @@
-export interface SpacingShape {
-  mobile?: string;
-  desktop: string;
-}
-
-export interface StyledSpacerProps {
-  $spacing?: 'policies' | 'docs' | 'simple' | 'normal';
-}
