--- src/components/sections/rich-text/themes/policies/index.tsx
@@ -1,9 +1,0 @@
-import { StyledPoliciesUL } from './policies.styles';
-import { defaultComponents } from '../default';
-
-export const policiesComponents = {
-  ...defaultComponents,
-  ul({ children }: Component) {
-    return <StyledPoliciesUL>{children}</StyledPoliciesUL>;
-  },
-};
