--- src/components/sections/rich-text/themes/policies/policies.styles.ts
@@ -1,8 +1,0 @@
-'use client';
-
-import { styled } from '@/styles';
-
-export const StyledPoliciesUL = styled.ul`
-  list-style: disc;
-  padding-left: 20px;
-`;
